Update tests to reflect changes to session.ts
