+++ webpack/session_keys.ts
@@ -1,0 +1,12 @@
+export enum BooleanSetting {
+  X_AXIS_INVERTED = "x_axis_inverted",
+  Y_AXIS_INVERTED = "y_axis_inverted",
+  Z_AXIS_INVERTED = "z_axis_inverted",
+  RAW_ENCODERS = "raw_encoders",
+  SCALED_ENCODERS = "scaled_encoders"
+}
+
+export enum NumericSetting {
+  BOT_ORIGIN_QUADRANT = "bot_origin_quadrant",
+  ZOOM_LEVEL = "zoom_level"
+}
